+++ analysis/__init__.py
@@ -2,6 +2,8 @@
 
 from .walk_forward import WalkForwardAnalyzer, WalkForwardResult, generate_adaptive_windows
 from .monte_carlo import MonteCarloSimulator, MonteCarloResult, print_mc_summary
+from .strategy_ranker import StrategyRanker, RankedStrategy, RankingConfig, rank_strategies
+from .visualizer import StrategyVisualizer, VisualizationConfig, visualize_backtest
 
 __all__ = [
     "WalkForwardAnalyzer", 
@@ -10,4 +12,13 @@ __all__ = [
     "MonteCarloSimulator",
     "MonteCarloResult",
     "print_mc_summary",
+    "StrategyRanker",
+    "RankedStrategy",
+    "RankingConfig",
+    "rank_strategies",
+    "StrategyVisualizer",
+    "VisualizationConfig",
+    "visualize_backtest",
 ]
+
+